# swagger.json
{
  "swagger": "2.0",
  "info": {
     "description": "A global Company that make travel and accommodation easy and convenient for the strong workforce of savvy members of staff, by leveraging the modern web.",
     "version": "1.0.0",
     "title": "firestar-backend",
     "termsOfService": "http://swagger.io/terms/",
     "contact": {
        "email": "apiteam@swagger.io"
     },
     "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
     }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
     "http"
  ],
  "paths": {
...
     "/user/{username}": {
        "get": {
           "tags": [
              "user"
           ],
           "summary": "Get user by user name",
           "description": "",
           "operationId": "getUserByName",
           "produces": [
              "application/xml",
              "application/json"
           ],
           "parameters": [
              {
                 "name": "username",
                 "in": "path",
                 "description": "The name that needs to be fetched. Use user1 for testing. ",
                 "required": true,
                 "type": "string"
              }
           ],
           "responses": {
              "200": {
                 "description": "successful operation",
                 "schema": {
                    "$ref": "#/definitions/User"
                 }
              },
              "400": {
                 "description": "Invalid username supplied"
              },
              "404": {
                 "description": "User not found"
              }
           }
        },
        "put": {
           "tags": [
              "user"
           ],
           "summary": "Updated user",
           "description": "This can only be done by the logged in user.",
           "operationId": "updateUser",
           "produces": [
              "application/xml",
              "application/json"
           ],
           "parameters": [
              {
                 "name": "username",
                 "in": "path",
                 "description": "name that need to be updated",
                 "required": true,
                 "type": "string"
              },
              {
                 "in": "body",
                 "name": "body",
                 "description": "Updated user object",
                 "required": true,
                 "schema": {
                    "$ref": "#/definitions/User"
                 }
              }
           ],
           "responses": {
              "400": {
                 "description": "Invalid user supplied"
              },
              "404": {
                 "description": "User not found"
              }
           }
        },
        "delete": {
           "tags": [
              "user"
           ],
           "summary": "Delete user",
           "description": "This can only be done by the logged in user.",
           "operationId": "deleteUser",
           "produces": [
              "application/xml",
              "application/json"
           ],
           "parameters": [
              {
                 "name": "username",
                 "in": "path",
                 "description": "The name that needs to be deleted",
                 "required": true,
                 "type": "string"
              }
           ],
           "responses": {
              "400": {
                 "description": "Invalid username supplied"
              },
              "404": {
                 "description": "User not found"
              }
           }
        }
     }
  },
  "securityDefinitions": {
     "petstore_auth": {
        "type": "oauth2",
        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
        "flow": "implicit",
        "scopes": {
           "write:pets": "modify pets in your account",
           "read:pets": "read your pets"
        }
     },
     "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
     }
  },
  "definitions": {
...
     "User": {
        "type": "object",
        "properties": {
           "id": {
              "type": "integer",
              "format": "int64"
           },
           "username": {
              "type": "string"
           },
           "firstName": {
              "type": "string"
           },
           "lastName": {
              "type": "string"
           },
           "email": {
              "type": "string"
           },
           "password": {
              "type": "string"
           },
           "phone": {
              "type": "string"
           },
           "userStatus": {
              "type": "integer",
              "format": "int32",
              "description": "User Status"
           }
        },
        "xml": {
           "name": "User"
        }
     },
...
  },
...
}